+++ main.js
@@ -2,6 +2,7 @@ import './module/settings.js';
 import './module/menu.js';
 import './module/token.js';
 import './module/party-cluster.js';
+import './module/cult.js';
 import './module/scene-music.js';
 import { migrate } from './module/migrate.js';
 import('./module/encounter.js');